requirements.txt
numpy>=2.1
pandas>=2.2
scikit-learn>=1.6
scipy>=1.14
joblib>=1.4
mlflow>=2.17
dvc>=3.59
prometheus-client>=0.21
fastapi>=0.115
uvicorn>=0.32
streamlit>=1.40
requests>=2.32
pyyaml>=6.0
